chore: created ModelPropertyHelper
